Getting close to release.
Did a bunch of experiments for the deadzone and I'm going to just turn it off for now. There is a path to maybe fixing the bug, and I was also looking for a way to manipulate the angle at which the player repositions without success. In that regard I can think of one more possible solution. But for now, the current features are pretty polished. A few more bits and bobs I think.